# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "DeleteResponse": {
        "properties": {
          "deleted": {
            "title": "Deleted",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "deleted",
          "id"
        ],
        "title": "DeleteResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobCostsBreakdownItem": {
        "description": "Sprint J.3.b — one row of the per-(provider, model) cost breakdown.",
        "properties": {
          "calls": {
            "title": "Calls",
            "type": "integer"
          },
          "completion_tokens": {
            "title": "Completion Tokens",
            "type": "integer"
          },
          "cost_usd": {
            "title": "Cost Usd",
            "type": "number"
          },
          "latency_ms": {
            "title": "Latency Ms",
            "type": "integer"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "prompt_tokens": {
            "title": "Prompt Tokens",
            "type": "integer"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "model",
          "calls",
          "cost_usd",
          "prompt_tokens",
          "completion_tokens",
          "latency_ms"
        ],
        "title": "JobCostsBreakdownItem",
        "type": "object"
      },
      "JobCostsKindItem": {
        "description": "§17.90 — one row of the per-call_kind cost breakdown.\n\n``kind`` is the literal string from ``llm_call_logs.call_kind``\nwith NULL folded into the sentinel ``\"uncategorized\"`` so consumers\ndon't have to handle NULL. Currently only ``\"synthesis\"`` is set\nexplicitly (by ``_synthesize_compiled_output``); everything else\nlands in ``\"uncategorized\"``.",
        "properties": {
          "calls": {
            "title": "Calls",
            "type": "integer"
          },
          "completion_tokens": {
            "title": "Completion Tokens",
            "type": "integer"
          },
          "cost_usd": {
            "title": "Cost Usd",
            "type": "number"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "latency_ms": {
            "title": "Latency Ms",
            "type": "integer"
          },
          "prompt_tokens": {
            "title": "Prompt Tokens",
            "type": "integer"
          }
        },
        "required": [
          "kind",
          "calls",
          "cost_usd",
          "prompt_tokens",
          "completion_tokens",
          "latency_ms"
        ],
        "title": "JobCostsKindItem",
        "type": "object"
      },
      "JobCostsResponse": {
        "description": "Sprint J.3.b — aggregate cost + latency for one job, with breakdown.\n\n``by_provider`` is sorted descending by cost_usd then calls so the\nbiggest spend lines surface first. ``call_count``/``total_*`` are\njob-wide totals across all (provider, model) combinations.\n\n§17.90 added ``by_kind`` — same shape, grouped by ``call_kind``.\nUseful for splitting compile-time synthesis spend from execution\nspend (W.7 follow-up).",
        "properties": {
          "by_kind": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/JobCostsKindItem"
            },
            "title": "By Kind",
            "type": "array"
          },
          "by_provider": {
            "items": {
              "$ref": "#/components/schemas/JobCostsBreakdownItem"
            },
            "title": "By Provider",
            "type": "array"
          },
          "call_count": {
            "title": "Call Count",
            "type": "integer"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "total_completion_tokens": {
            "title": "Total Completion Tokens",
            "type": "integer"
          },
          "total_cost_usd": {
            "title": "Total Cost Usd",
            "type": "number"
          },
          "total_latency_ms": {
            "title": "Total Latency Ms",
            "type": "integer"
          },
          "total_prompt_tokens": {
            "title": "Total Prompt Tokens",
            "type": "integer"
          }
        },
        "required": [
          "job_id",
          "total_cost_usd",
          "total_prompt_tokens",
          "total_completion_tokens",
          "total_latency_ms",
          "call_count",
          "by_provider"
        ],
        "title": "JobCostsResponse",
        "type": "object"
      },
      "JobListResponse": {
        "properties": {
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/app__schemas__JobSummary"
            },
            "title": "Jobs",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "jobs",
          "total",
          "limit",
          "offset"
        ],
        "title": "JobListResponse",
        "type": "object"
      },
      "JobRenameInput": {
        "properties": {
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "title": "JobRenameInput",
        "type": "object"
      },
      "JobSynthesisOverrideInput": {
        "description": "Sprint X.6 — per-job opt-in for the W.7 LLM synthesis pass.\n\nNone inherits ``settings.compile_synthesis_enabled``; True forces\nsynthesis on for this job; False forces it off. The field is\nexplicitly Optional[bool] (not ``bool | None`` with a default) so\nconsumers MUST declare intent — sending an absent body would be\nindistinguishable from \"set to null\".",
        "properties": {
          "override": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Override"
          }
        },
        "required": [
          "override"
        ],
        "title": "JobSynthesisOverrideInput",
        "type": "object"
      },
      "JobSynthesisOverrideResponse": {
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "override": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Override"
          }
        },
        "required": [
          "job_id",
          "override"
        ],
        "title": "JobSynthesisOverrideResponse",
        "type": "object"
      },
...
      "ResumeJobInput": {
        "description": "Body for POST /jobs/{job_id}/resume. ``job_id`` comes from the path.",
        "properties": {
          "model_overrides": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Overrides"
          },
          "skip_optimize": {
            "default": false,
            "title": "Skip Optimize",
            "type": "boolean"
          },
          "skip_verify": {
            "default": false,
            "title": "Skip Verify",
            "type": "boolean"
          }
        },
        "title": "ResumeJobInput",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "app__schemas__JobSummary": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "node_count": {
            "default": 0,
            "title": "Node Count",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "JobSummary",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Self-hosted RAG-powered workflow orchestrator",
    "title": "Scaffold Engine",
    "version": "1.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/jobs": {
      "get": {
        "description": "Paginated job list with optional status filter and title search.\n\nSprint X.9 — ``synthesized`` filter complements the X.6 per-job opt-in:\n``?synthesized=true`` lists only jobs whose ``compiled_output`` was\nLLM-synthesized (W.7 narrative pass); ``?synthesized=false`` lists\neverything else (heuristic compile, unsynthesized, or not-yet-compiled).\nOmit the param to see all jobs.",
        "operationId": "list_jobs_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "synthesized",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Synthesized"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List Jobs",
        "tags": [
          "Management"
        ]
      }
    },
    "/jobs/cleanup": {
      "post": {
        "description": "Find and resolve stale/orphaned jobs. Requires API key (global auth).",
        "operationId": "cleanup_stale_jobs_jobs_cleanup_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Cleanup Stale Jobs",
        "tags": [
          "ops"
        ]
      }
    },
    "/jobs/{job_id}": {
      "delete": {
        "description": "Hard-delete a job. Cascade removes dag_nodes / execution_logs / artifacts /\nerror_logs (FK ON DELETE CASCADE). llm_call_logs rows are unaffected\n(no FK; off-job calls live there too).",
        "operationId": "delete_job_jobs__job_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Delete Job",
        "tags": [
          "Management"
        ]
      },
      "patch": {
        "description": "Rename a job (set title).",
        "operationId": "rename_job_jobs__job_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobRenameInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__schemas__JobSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Rename Job",
        "tags": [
          "Management"
        ]
      }
    },
    "/jobs/{job_id}/costs": {
      "get": {
        "description": "Sprint J.3.b — aggregate cost + latency for a job.\n\nReturns total USD spent (computed at insert time, immutable),\ntotal prompt/completion tokens, total LLM latency, the count of\nLLM calls logged for this job, and a per-(provider, model)\nbreakdown sorted by descending cost. Job_ids with no logged\ncalls return the zero shape with an empty breakdown — fail-open\nmatches the rest of the cost-tracking surface.",
        "operationId": "get_job_costs_endpoint_jobs__job_id__costs_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobCostsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Job Costs Endpoint",
        "tags": [
          "Management"
        ]
      }
    },
    "/jobs/{job_id}/resume": {
      "post": {
        "description": "Resume a cancelled job and stream its execution.\n\nAtomically flips the job from ``cancelled`` back to ``executing`` and\nre-fires ``/execute/all``-equivalent execution. Replaces the manual\nSQL-then-curl recipe in debugging.md. ``execute_all_nodes`` is\nidempotent over completed nodes — execution picks up from the last\npending node, with done-node outputs serving as upstream context.\n\nStatus codes:\n  - 200 + SSE stream on successful resume\n  - 404 if no job with that ID exists\n  - 409 if the job exists but isn't in ``cancelled`` (current status\n    returned in detail for client-side dispatch)\n  - 400 on malformed UUID",
        "operationId": "resume_job_endpoint_jobs__job_id__resume_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeJobInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Resume Job Endpoint",
        "tags": [
          "Management"
        ]
      }
    },
    "/jobs/{job_id}/synthesis": {
      "patch": {
        "description": "Sprint X.6 — set the per-job opt-in for the W.7 LLM synthesis pass.\n\nBody ``{\"override\": true}`` forces synthesis on for this job;\n``{\"override\": false}`` forces it off; ``{\"override\": null}`` clears\nthe override so the job inherits ``settings.compile_synthesis_enabled``.\n\nThe override is read by ``execution_compile._resolve_synthesis_enabled``\non the next compile pass — set it before ``/execute/all`` (or before\na final-node retry) for it to take effect on the resulting deliverable.",
        "operationId": "set_job_synthesis_override_jobs__job_id__synthesis_patch",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobSynthesisOverrideInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSynthesisOverrideResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Set Job Synthesis Override",
        "tags": [
          "Management"
        ]
      }
    },
...
    "/observability/jobs": {
      "get": {
        "description": "Recent jobs with their LLM cost + latency totals.\n\nLists jobs created within the window, sorted by total cost DESC\n(then created_at DESC). Each row carries the job's status and its\naggregated cost/tokens/latency from ``llm_call_logs``. Jobs with\nzero logged calls (planning-only, or pre-J.3.a) appear with zeros\nrather than being omitted.\n\nUseful for \"what's been expensive in the last hour/day\" without\npaging through ``/jobs`` and calling ``/jobs/{id}/costs`` per row.",
        "operationId": "recent_jobs_endpoint_observability_jobs_get",
        "parameters": [
          {
            "description": "Job creation window in minutes (default 24h).",
            "in": "query",
            "name": "window_minutes",
            "required": false,
            "schema": {
              "default": 1440,
              "description": "Job creation window in minutes (default 24h).",
              "maximum": 10080,
              "minimum": 1,
              "title": "Window Minutes",
              "type": "integer"
            }
          },
          {
            "description": "Max rows returned (default 25).",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "description": "Max rows returned (default 25).",
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Recent Jobs Endpoint Observability Jobs Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Recent Jobs Endpoint",
        "tags": [
          "Observability"
        ]
      }
    },
...
  }
}